TriggerZone Conditionals/Follower Interaction
- Added Conditions to the TriggerZones.
now any key in the gamestats dictionary can be tested to see if its
there or not and triggerzones will olny trigger when ALL conditions are
met. The triggerzones can also have as many of these conditions as it
wants. the delimiter is a space.
- Moved all the Mirror stuff off to the side and disabled it until i
decide what  to do with it.
- Started Adding in the functionality for the Follower to interact with
the
player through the use of the FSM and the TriggerZones.