Npc Animation
* Added the <Sprite> tag to the NpcXml it controls the Npcs Animations.
* Added a HasSprite bool to the Npc Family along with a SpriteDir and
SpriteName strings for the animation and the code to call there idle and
walking animations.
* Fixed The Pathfinding For The Npcs. They were going right through the
solids that were created at runtime so i added a some events to the
object spawner to fix it.
* theres still some bug in the movement code where they still go through
the obsticals i thinks it has to do with the custom movement missing the
node and drifting out forever.